feat: reorganize template processing under notifications package
