+++ openmetadata-service/src/main/java/org/openmetadata/service/jdbi3/NotificationTemplateRepository.java
@@ -25,9 +25,9 @@ import org.openmetadata.schema.type.Include;
 import org.openmetadata.schema.type.ProviderType;
 import org.openmetadata.schema.type.change.ChangeSource;
 import org.openmetadata.service.Entity;
+import org.openmetadata.service.notifications.template.NotificationTemplateProcessor;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsNotificationTemplateProcessor;
 import org.openmetadata.service.resources.events.NotificationTemplateResource;
-import org.openmetadata.service.template.NotificationTemplateProcessor;
-import org.openmetadata.service.template.handlebars.HandlebarsNotificationTemplateProcessor;
 import org.openmetadata.service.util.EntityUtil;
 import org.openmetadata.service.util.EntityUtil.Fields;
 import org.openmetadata.service.util.resourcepath.ResourcePathResolver;
