+++ openmetadata-service/src/main/java/org/openmetadata/service/notifications/template/handlebars/helpers/DiffHelper.java
@@ -1,10 +1,10 @@
-package org.openmetadata.service.template.handlebars.helpers;
+package org.openmetadata.service.notifications.template.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import com.github.jknack.handlebars.Options;
 import java.util.LinkedList;
 import org.bitbucket.cowwoc.diffmatchpatch.DiffMatchPatch;
-import org.openmetadata.service.template.handlebars.HandlebarsHelper;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsHelper;
 
 /**
  * Helper to generate inline diff with add/remove markers. Usage: {{diff oldValue newValue}}
