+++ openmetadata-service/src/main/java/org/openmetadata/service/notifications/template/handlebars/helpers/EqHelper.java
@@ -1,8 +1,8 @@
-package org.openmetadata.service.template.handlebars.helpers;
+package org.openmetadata.service.notifications.template.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import java.util.Objects;
-import org.openmetadata.service.template.handlebars.HandlebarsHelper;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsHelper;
 
 /**
  * Helper to check if two values are equal.
