+++ openmetadata-service/src/main/java/org/openmetadata/service/notifications/template/handlebars/helpers/FormatColumnValueHelper.java
@@ -1,8 +1,8 @@
-package org.openmetadata.service.template.handlebars.helpers;
+package org.openmetadata.service.notifications.template.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import java.util.Map;
-import org.openmetadata.service.template.handlebars.HandlebarsHelper;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsHelper;
 
 /**
  * Helper to format column values based on their type. Usage: {{formatColumnValue columnData}}
