+++ openmetadata-service/src/main/java/org/openmetadata/service/notifications/template/handlebars/helpers/ResolveDomainHelper.java
@@ -1,4 +1,4 @@
-package org.openmetadata.service.template.handlebars.helpers;
+package org.openmetadata.service.notifications.template.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import java.util.Map;
@@ -6,7 +6,7 @@ import java.util.UUID;
 import org.openmetadata.schema.entity.domains.Domain;
 import org.openmetadata.schema.type.Include;
 import org.openmetadata.service.Entity;
-import org.openmetadata.service.template.handlebars.HandlebarsHelper;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
