+++ openmetadata-service/src/main/java/org/openmetadata/service/notifications/template/handlebars/helpers/SplitHelper.java
@@ -1,8 +1,8 @@
-package org.openmetadata.service.template.handlebars.helpers;
+package org.openmetadata.service.notifications.template.handlebars.helpers;
 
 import com.github.jknack.handlebars.Handlebars;
 import java.util.regex.Pattern;
-import org.openmetadata.service.template.handlebars.HandlebarsHelper;
+import org.openmetadata.service.notifications.template.handlebars.HandlebarsHelper;
 
 /**
  * Helper to split a string by a delimiter.
